requirements.txt
transformers
transformers[textsum]
transformers[sentpiece]
datasets
sacrebleu 
rouge_score 
py7zr
pandas
nltk
tqdm
PyYAML
matplotlib
torch
pocketsphinx
notebook
boto3
tensorflow
pandas
mypy-boto3-s3
python-box==6.0.2
ensure==1.0.2
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2
#-e .
